Update how-to-style.md
i feel this document completely misses customizing toast messages differently.  more examples are needed as this took me a long time to get to work as I wanted with no examples other than global styling with ToastContainer

"- className: applied on the toast wrapper (this override `toastClassName` is set by the container )"  does this replace the entire container styling and i have to do it all myself or does it just change the background color?  I think a lot of the suggestions above are very confusing without concrete full examples

thanks!